fix pyright errors

+++ src/bidirectional.py
@@ -24,7 +24,7 @@ class BiDirExp:
     sol_nvars: int
     sol_nhard: int
     sol_nsoft: int
-    sol_navgclause: int
+    sol_navgclause: float
     sol_ntotalvars: int
     sol_nmaxclause: int
     factor_size: int
@@ -58,7 +58,7 @@ class BiDirExp:
       sol_nhard =                    0,
       sol_nsoft =                    0,
       factor_size =                  0,
-      sol_navgclause =               0,
+      sol_navgclause =               0.0,
       sol_ntotalvars =               0,
       sol_nmaxclause =               0,
       factors =                      BiDirType([])
